fix issue with freezing red5 server

+++ apps_v1.5/puppet/lib/wkm/client.js
@@ -16,6 +16,7 @@ module.exports = structr({
 			this.start([options.rtmp.protocol, "//", options.rtmp.hostname, ":", options.rtmp.port, options.rtmp.path].join(""));
 		}
 		this._width = this._height = 0;
+		this._killing = false;
 	},
 
 	/**
@@ -31,30 +32,42 @@ module.exports = structr({
 
 	'start': function(rtmpUrl) {
 
-		console.log("broadcasting to rtmp server %s", rtmpUrl);
 		var self = this;
+		// rtmpUrl = "C:\\Users\\Administrator\\Desktop\\test.flv"
 
-		this._rtmpUrl = rtmpUrl;
+		if(arguments.length) this._rtmpUrl = rtmpUrl || this._rtmpUrl;
 
-		this.kill();
 
-		this._proc = spawn(__dirname + "/../../../wkm/bin/cli.exe", [rtmpUrl]);
 
-		this._proc.stdout.on('data', function(data) {
-			process.stdout.write(String(data));
+		this.kill(function() {
+			console.log("broadcasting to rtmp server %s", self._rtmpUrl);
+			self._proc = spawn(__dirname + "/../../../wkm/bin/cli.exe", [self._rtmpUrl, 
+				self._width, 
+				self._height, 
+				self._padding[0], 
+				self._padding[1],
+				self._padding[2],
+				self._padding[3]]);
+
+			self._proc.stdout.on('data', function(data) {
+				process.stdout.write(String(data));
+			});
+
+			self._proc.stderr.on('data', function(data) {
+				process.stderr.write(String(data));
+			});
+
+			self._proc.on("exit", function() {
+				self._proc = null;
+				console.log("EXIT")
+			});
 		});
 
-		this._proc.stderr.on('data', function(data) {
-			process.stderr.write(String(data));
-		});
-
-		this._proc.on("exit", function() {
-			self._proc = null;
-		})
+		
 
 
-		this.padding(this._padding[0], this._padding[1], this._padding[2], this._padding[3]);
-		this._resize(this._width, this._height);
+		//this.padding(this._padding[0], this._padding[1], this._padding[2], this._padding[3]);
+		//this._resize(this._width, this._height);
 
 		return this;
 	},
@@ -63,10 +76,21 @@ module.exports = structr({
 	 * kills WKM
 	 */
 
-	'kill': function() {
+	'kill': function(callback) {
+		var self = this;
+		if(this._killing) return;
+		this._killing = true;
+		if(!callback) callback = function(){};
 		if(this._proc) {
+			console.log("KILL")
+			this._proc.once("exit", function() {
+				self._killing = false;
+				callback();
+			});
 			this._proc.kill();
 			this._proc = undefined;
+		} else {
+			callback();
 		}
 	},
 
@@ -120,18 +144,12 @@ module.exports = structr({
 
 	'padding': function(padding) {
 		this._padding = this._args(padding.left, padding.right, padding.top, padding.bottom);
-		if(this._proc) this._proc.stdin.write('padding\n' + this._padding.join(' ') + "\n");
+		if(!this._proc) return;
+		 // this._proc.stdin.write('padding\n' + this._padding.join(' ') + "\n");
 		//resize here???
+		this.resizeDesktop();
 	},
 
-	/**
-	 */
-
-	"_resize": function(width, height) {
-		// this._width = width || this._width;
-		// this._height = height || this._height;
-		if(this._proc) this._proc.stdin.write('resize\n' + this._args(this._width, this._height).join(" ") + "\n");
-	},
 
 	/**
 	 */
@@ -145,7 +163,7 @@ module.exports = structr({
 			process.stdout.write(stdout);
 			process.stderr.write(stderr);
 		});
-		this._resize(width, height);
+		this.start(); //restart
 	},
 
 	/**
